% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{county_fips}
\alias{county_fips}
\title{Provides geographic information about the inputted county FIPS code}
\usage{
county_fips(county_fips)
}
\arguments{
\item{county_fips}{The county FIPS code that you want info about.}
}
\value{
A dataframe with the zipcode, city, county, state, and state FIPS code
corresponding to the inputted county FIPS code.
}
\description{
Provides geographic information about the inputted county FIPS code
}
\examples{
county_fips("01001")
county_fips(01001)

county_fips(c("05043", "12033", "06065"))
county_fips(c(05043, 12033, 06065))
}

