% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{state_fips}
\alias{state_fips}
\title{Provides geographic information about the inputted state FIPS code}
\usage{
state_fips(state_fips)
}
\arguments{
\item{state_fips}{State FIPS code or codes that you would like information about}
}
\value{
A dataframe with the zipcode, city, county, state, and county FIPS code
corresponding to the inputted state FIPS code.
}
\description{
Provides geographic information about the inputted state FIPS code
}
\examples{
state_fips("01")
state_fips(01)

state_fips(c("05", "25", "11"))
state_fips(c(05, 25, 11))
}

