% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{zipcode}
\alias{zipcode}
\title{Provides geographic information about the inputted zipcodes.}
\usage{
zipcode(zipcode)
}
\arguments{
\item{zipcode}{Zipcode or zipcodes you would like information about}
}
\value{
A dataframe with the city, county, state, county FIPS code, and state
FIPS code corresponding to the inputted zipcode.
}
\description{
Provides geographic information about the inputted zipcodes.
}
\examples{

zipcode("19104")
zipcode(19104)

zipcode(c(19104, 19105))

example <- c(19104, 36067, 36003)
zipcode(example)
}

