% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{zipcodes}
\alias{zipcodes}
\title{City, county, state, county FIP code and state FIP code for 40569 zipcodes
in the United States}
\format{A data frame with 40569 rows and 6 variables:
\describe{
  \item{zipcode}{Zipcode number}
  \item{city}{City name}
  \item{county}{County name}
  \item{state}{State name}
  \item{county_FIPS}{FIP code for the county}
  \item{state_FIPS}{FIP code for the state}
  ...
}}
\source{
\url{zipcodestogo.com}
}
\usage{
zipcodes
}
\description{
A dataset containing the city, county, state, county FIPS code, and state
FIPS code for 40569 zipcodes in the United States
}
\keyword{datasets}

