% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{zipcodes_city}
\alias{zipcodes_city}
\title{Provides geographic information about the inputted city}
\usage{
zipcodes_city(city, county = FALSE, state = FALSE)
}
\arguments{
\item{city}{The city name that you are interested in.}

\item{county}{Optional county to be more specific in your queries.}

\item{state}{Optional state to be more specific in your queries.}
}
\value{
A dataframe with the zipcode, county, state, county FIPS codes,
and state FIPS code corresponding to the inputted city.
}
\description{
Provides geographic information about the inputted city
}
\examples{
zipcodes_city(city = "Sacramento")
zipcodes_city(city = "Sacramento", county = "Sacramento")
zipcodes_city(city = "Sacramento", state = "California")
zipcodes_city(city = "Sacramento", county = "Sacramento",
              state = "California")

zipcodes_city(city = c("Sacramento", "Walnut Creek"))
}

