% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{zipcodes_county}
\alias{zipcodes_county}
\title{Provides geographic information about the inputted county}
\usage{
zipcodes_county(county, state = FALSE)
}
\arguments{
\item{county}{The county name or names that you are interested in.}

\item{state}{Optional state to be more specific in your queries.}
}
\value{
A dataframe with the zipcode, city, state, county FIPS codes,
and state FIPS code corresponding to the inputted county.
}
\description{
Provides geographic information about the inputted county
}
\examples{
zipcodes_county("sacramento")
zipcodes_county(county = "sacramento", state = "California")
zipcodes_county(county = c("sacramento", "Alameda"), state = "California")
}

