% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{zipcodes_state}
\alias{zipcodes_state}
\title{Provides geographic information about the inputted state}
\usage{
zipcodes_state(states)
}
\arguments{
\item{states}{The states that you would like info about.}
}
\value{
A dataframe with the zipcode, city, county, county FIPS codes,
and state FIPS code corresponding to the inputted state.
}
\description{
Provides geographic information about the inputted state
}
\examples{
zipcodes_state("california")
zipcodes_state("California")

zipcodes_state(c("california", "wyoming"))
zipcodes_state(c("California", "Wyoming"))
}

