% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{zips2fips}
\alias{zips2fips}
\title{Converts zipcodes into county FIPS codes for those locations}
\usage{
zips2fips(zipcode)
}
\arguments{
\item{zipcode}{The zipcodes that you want the county FIPS codes for.}
}
\value{
A data.frame with one column, the county FIPS code that correspond
to the inputted zipcodes.
}
\description{
Converts zipcodes into county FIPS codes for those locations
}
\examples{
zips2fips(c("94526", "19104"))
}

